+++ src/ui/CommandListener.js
@@ -2,8 +2,6 @@
 "use strict";
 
 import { CompositeDisposable } from 'atom';
-const ElearnJSMenuManager = require('./ElearnJSMenuManager.js');
-const FileWriter = require('../export/FileWriter.js');
 
 class CommandListener {
     constructor(fileWriter, elearnJSMenuManager) {