+++ src/ui/OptionMenuManager.js
@@ -1,7 +1,5 @@
 "use strict";
 
-const { dialog } = require('electron').remote;
-
 const OptionMenu = require("./OptionMenu.js");
 
 class OptionMenuManager {